+++ hutool-log/src/main/java/cn/hutool/log/dialect/console/ConsoleLog.java
@@ -21,10 +21,20 @@ public class ConsoleLog extends AbstractLog {
 	private String name;
 	
 	//------------------------------------------------------------------------- Constructor
+	/**
+	 * 构造
+	 * 
+	 * @param clazz 类
+	 */
 	public ConsoleLog(Class<?> clazz) {
 		this.name = clazz.getName();
 	}
 	
+	/**
+	 * 构造
+	 * 
+	 * @param name 类名
+	 */
 	public ConsoleLog(String name) {
 		this.name = name;
 	}